refactor(oauth): extract Postgres atrium-oauth stores into atproto-oauth-pg-store crate
Move the PgStateStore/PgSessionStore atrium-oauth Store adapters (from
observing-appview) and their backing SQL (from observing-db::oauth) into a
new, self-contained crate. Queries are converted to runtime sqlx::query so
the crate compiles without a live DB or vendored .sqlx cache; the two
required tables ship as a migration. State TTL is now configurable
(PgStateStore::with_ttl_ms).

observing-appview and auth.rs now depend on the new crate; observing-db
drops its oauth module. Orphaned .sqlx entries for the moved queries are
removed so 'cargo sqlx prepare --check' stays green.
