--- .sqlx/query-7695b5c3a08490a39d2e252bf08b2f072809f67a630731ceb58f545c73816bca.json
@@ -1,14 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "DELETE FROM oauth_state WHERE key = $1",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "7695b5c3a08490a39d2e252bf08b2f072809f67a630731ceb58f545c73816bca"
-}
