--- .sqlx/query-a0f6a08da6a6433205588ce42cab3014e611075a5b2f38681f5f02f2a3ce3bea.json
@@ -1,22 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "SELECT value FROM oauth_sessions WHERE key = $1",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "value",
-        "type_info": "Text"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Text"
-      ]
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "a0f6a08da6a6433205588ce42cab3014e611075a5b2f38681f5f02f2a3ce3bea"
-}
