--- .sqlx/query-c74935f87043c76921389aa79e3d6a59522a414b569e0c5ccf734cb4371bef55.json
@@ -1,16 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n        INSERT INTO oauth_state (key, value, expires_at)\n        VALUES ($1, $2, NOW() + ($3 || ' milliseconds')::interval)\n        ON CONFLICT (key) DO UPDATE SET value = $2, expires_at = NOW() + ($3 || ' milliseconds')::interval\n        ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text",
-        "Text",
-        "Text"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "c74935f87043c76921389aa79e3d6a59522a414b569e0c5ccf734cb4371bef55"
-}
