--- .sqlx/query-d282892a66da0b07618c1443c5210e61e36d936871e5840197869b6b8d14ab43.json
@@ -1,22 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "SELECT value FROM oauth_state WHERE key = $1 AND expires_at > NOW()",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "value",
-        "type_info": "Text"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Text"
-      ]
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "d282892a66da0b07618c1443c5210e61e36d936871e5840197869b6b8d14ab43"
-}
