+++ crates/observing-appview/Cargo.toml
@@ -20,6 +20,8 @@ atproto-blob-resolver = { path = "../atproto-blob-resolver" }
 file-blob-cache = { path = "../file-blob-cache" }
 # Wire types for the species-id HTTP client
 observing-species-id-protocol = { path = "../observing-species-id-protocol" }
+# Postgres-backed OAuth state/session stores for atrium-oauth
+atproto-oauth-pg-store = { path = "../atproto-oauth-pg-store" }
 
 # AT Protocol types (for record construction)
 jacquard-common = "0.12.0-beta.2"
