+++ crates/observing-appview/src/main.rs
@@ -5,7 +5,6 @@ mod enrichment;
 mod error;
 mod media;
 mod middleware;
-mod oauth_store;
 mod resolver;
 mod responses;
 mod routes;
