+++ crates/observing-db/src/lib.rs
@@ -7,7 +7,6 @@ pub mod interactions;
 pub mod likes;
 pub mod migrate;
 pub mod notifications;
-pub mod oauth;
 pub mod occurrences;
 pub mod private_data;
 #[cfg(feature = "processing")]
